{"schema":"decodex/commit/1","summary":"Modularize frozen export stroke rasterization","authority":"manual"}
